# openapi.yaml
openapi: 3.0.3
info:
  title: Awesomeness Photography Backend API
  description: |
    A comprehensive photography portfolio and booking management system API built with Node.js, Express, TypeScript, AWS DynamoDB, and AWS S3.
    
    ## Features
    - Event Type Management (Wedding, Portrait, Corporate, etc.)
    - Event Portfolio Management
    - Image Upload and Storage (AWS S3)
    - Customer Booking System
    - Admin Authentication and Management
    
    ## Authentication
    Admin routes require an API key passed in the `x-api-key` header.
    
    ## Base URL
    - Development: `http://localhost:3000/api`
    - Production: `https://your-domain.com/api`
    
  version: 0.1.0
  contact:
    name: Awesomeness Backend Support
    email: support@example.com
  license:
    name: Private
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.awesomeness.com/api
    description: Production server
...
paths:
...
  /images/event/{eventId}:
    get:
      tags:
        - Images
      summary: List images by event
      description: Get all images for a specific event (public endpoint)
      operationId: listImagesByEvent
      parameters:
        - name: eventId
          in: path
          required: true
          description: Event ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Images retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /images:
    post:
      tags:
        - Images
      summary: Upload images (Admin)
      description: Upload one or multiple images to an event (admin only)
      operationId: uploadImages
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - images
                - eventId
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Image files to upload (max 10)
                  maxItems: 10
                eventId:
                  type: string
                  format: uuid
                  description: ID of the event to associate images with
                description:
                  type: string
                  description: Optional description for the images
      responses:
        '200':
          description: Images uploaded successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /bookings:
    post:
      tags:
        - Bookings
      summary: Create booking
      description: Create a new booking request (public endpoint)
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventTypeId
                - customerName
                - customerEmail
                - customerPhone
                - eventDate
              properties:
                eventTypeId:
                  type: string
                  format: uuid
                  example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                  description: ID of the event type for booking
                customerName:
                  type: string
                  example: John Doe
                  description: Customer's full name
                customerEmail:
                  type: string
                  format: email
                  example: john@example.com
                  description: Customer's email address
                customerPhone:
                  type: string
                  example: '+1234567890'
                  description: Customer's phone number
                eventDate:
                  type: string
                  format: date
                  example: '2025-12-25'
                  description: Desired event date
                eventLocation:
                  type: string
                  example: Central Park, New York
                  description: Event location
                message:
                  type: string
                  example: Looking for wedding photography services
                  description: Additional message or requirements
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequestError'

    get:
      tags:
        - Bookings
      summary: List bookings (Admin)
      description: Get all bookings, optionally filtered by event type (admin only)
      operationId: listBookings
      security:
        - ApiKeyAuth: []
      parameters:
        - name: eventTypeId
          in: query
          description: Filter bookings by event type ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Admin API key for authentication
  schemas:
...
    Image:
      type: object
      required:
        - id
        - eventId
        - s3Key
        - filename
        - contentType
        - size
        - uploadedAt
      properties:
        id:
          type: string
          format: uuid
          example: c3d4e5f6-a7b8-9012-cdef-1234567890ab
          description: Unique identifier for the image
        eventId:
          type: string
          format: uuid
          example: b2c3d4e5-f6a7-8901-bcde-f1234567890a
          description: ID of the associated event
        s3Key:
          type: string
          example: events/b2c3d4e5-f6a7-8901-bcde-f1234567890a/photo1.jpg
          description: S3 object key for the image
        filename:
          type: string
          example: photo1.jpg
          description: Original filename of the image
        contentType:
          type: string
          example: image/jpeg
          description: MIME type of the image
        size:
          type: integer
          example: 1024000
          description: File size in bytes
        description:
          type: string
          example: Ceremony photo
          description: Optional description of the image
          nullable: true
        uploadedAt:
          type: string
          format: date-time
          example: '2025-11-15T10:00:00.000Z'
          description: Timestamp when the image was uploaded
...
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - customerName
        - customerEmail
        - customerPhone
        - eventDate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: d4e5f6a7-b8c9-0123-def0-1234567890bc
          description: Unique identifier for the booking
        eventTypeId:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
          description: ID of the event type being booked
        customerName:
          type: string
          example: John Doe
          description: Customer's full name
        customerEmail:
          type: string
          format: email
          example: john@example.com
          description: Customer's email address
        customerPhone:
          type: string
          example: '+1234567890'
          description: Customer's phone number
        eventDate:
          type: string
          format: date
          example: '2025-12-25'
          description: Desired event date
        eventLocation:
          type: string
          example: Central Park, New York
          description: Event location
          nullable: true
        message:
          type: string
          example: Looking for wedding photography services
          description: Additional message or requirements
          nullable: true
        status:
          type: string
          enum:
            - pending
            - confirmed
            - cancelled
          example: pending
          description: Status of the booking
        createdAt:
          type: string
          format: date-time
          example: '2025-11-15T10:00:00.000Z'
          description: Timestamp when the booking was created
        updatedAt:
          type: string
          format: date-time
          example: '2025-11-15T10:00:00.000Z'
          description: Timestamp when the booking was last updated
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        message:
          type: string
          example: Detailed error message
  responses:
    BadRequestError:
      description: Bad request - Invalid input or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Invalid input parameters
    UnauthorizedError:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized - Invalid API key
    NotFoundError:
      description: Not found - The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
...